refactor: only render ChangeSumModal contents when open
